refactor: :truck: ファイルの格納箇所を変更
