+++ lib/core/page/profile_edit_page.dart
@@ -5,11 +5,11 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
 import '../../feature/user_profile/application/user_profile_for_write_notifier.dart';
+import '../../feature/user_profile/presentation/avatar_network_image_widget.dart';
 import '../../feature/user_profile/presentation/changeable_profile_image.dart';
 import '../../util/logger.dart';
 import '../../util/mixin/presentation_mixin.dart';
 import '../common_provider/dialog_controller.dart';
-import '../common_widget/avatar_network_image_widget.dart';
 import '../common_widget/dialog/confirm_dialog.dart';
 import '../common_widget/error_and_retry_widget.dart';
 import '../router/app_router.dart';
