+++ lib/core/page/word_top_page.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
+import '../../feature/definition/presentation/post_definition_fab.dart';
 import '../../feature/definition_list/presentation/definition_list.dart';
 import '../../feature/definition_list/util/definition_feed_type.dart';
 import '../../feature/word/application/word_state.dart';
@@ -11,7 +12,6 @@ import '../../feature/word/presentation/word_page_shimmer.dart';
 import '../../feature/word/presentation/word_widget.dart';
 import '../../util/extension/scroll_controller_extension.dart';
 import '../../util/logger.dart';
-import '../common_widget/button/post_definition_fab.dart';
 import '../common_widget/error_and_retry_widget.dart';
 import '../common_widget/stickey_tab_bar_deligate.dart';
