+++ lib/feature/definition/presentation/definition_tile.dart
@@ -5,12 +5,12 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
 import '../../../../core/common_widget/adaptive_overflow_text.dart';
-import '../../../../core/common_widget/avatar_network_image_widget.dart';
 import '../../../../core/common_widget/error_and_retry_widget.dart';
 import '../../../../core/router/app_router.dart';
 import '../../../../util/extension/date_time_extension.dart';
 import '../../../../util/logger.dart';
 import '../../definition_like/presentation/like_widget.dart';
+import '../../user_profile/presentation/avatar_network_image_widget.dart';
 import '../application/definition_state.dart';
 import 'definition_tile_shimmer.dart';
