+++ lib/feature/definition/presentation/post_definition_fab.dart
@@ -2,8 +2,8 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import '../../../feature/definition/presentation/write_definition_base_page.dart';
-import '../../router/app_router.dart';
+import '../../../core/router/app_router.dart';
+import 'write_definition_base_page.dart';
 
 /// 定義投稿画面へ遷移するFAB
 class PostDefinitionFAB extends StatelessWidget {
