+++ lib/feature/definition_list/presentation/individual_dictionary_definition_list.dart
@@ -5,11 +5,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
-import '../../../core/common_widget/button/post_definition_fab.dart';
 import '../../../core/common_widget/simple_widget_for_empty.dart';
 import '../../../util/constant/initial_main_group.dart';
 import '../../../util/extension/scroll_controller_extension.dart';
 import '../../auth/application/auth_state.dart';
+import '../../definition/presentation/post_definition_fab.dart';
 import '../../user_profile/presentation/dictionary_author_widget.dart';
 import '../util/definition_feed_type.dart';
 import 'definition_list.dart';
