+++ lib/feature/user_profile/presentation/changeable_profile_image.dart
@@ -6,10 +6,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:pull_down_button/pull_down_button.dart';
 
-import '../../../core/common_widget/avatar_network_image_widget.dart';
 import '../../../util/mixin/presentation_mixin.dart';
 import '../application/user_profile_for_write_notifier.dart';
 import '../domain/user_profile.dart';
+import 'avatar_network_image_widget.dart';
 
 class ChangeableProfileImage extends ConsumerWidget with PresentationMixin {
   const ChangeableProfileImage({
