+++ lib/feature/user_profile/presentation/profile_widget.dart
@@ -4,7 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
 
-import '../../../../../core/common_widget/avatar_network_image_widget.dart';
 import '../../../../../core/common_widget/button/outlined_button.dart';
 import '../../../../../core/common_widget/error_and_retry_widget.dart';
 import '../../../../../core/router/app_router.dart';
@@ -13,6 +12,7 @@ import '../../auth/application/auth_state.dart';
 import '../../user_follow/presentation/follow_or_unfollow_button.dart';
 import '../../user_follow/presentation/following_and_follower_count_widget.dart';
 import '../application/user_profile_state.dart';
+import 'avatar_network_image_widget.dart';
 import 'profile_widget_shimmer.dart';
 
 class ProfileWidget extends ConsumerWidget {
